Rename `NotablePersonBodySkeleton` to `NotablePersonBody`
