+++ src/app/pages/NotablePerson/LoadableNotablePerson.tsx
@@ -1,11 +1,11 @@
 import universal from 'react-universal-component';
 import { Props } from './NotablePerson';
-import { NotablePersonBodySkeleton } from './NotablePersonBodySkeleton';
+import { NotablePersonBody } from './NotablePersonBody';
 
 export const LoadableNotablePerson = universal<Props>(
   import('./NotablePerson'),
   {
     key: module => module.NotablePerson,
-    loading: NotablePersonBodySkeleton,
+    loading: NotablePersonBody,
   },
 );
