--- src/app/pages/NotablePerson/NotablePersonBodySkeleton.module.scss
@@ -1,5 +1,0 @@
-.root {
-  max-width: 100%;
-  display: flex;
-  flex-direction: column;
-}
